Start MLIR module
