+++ exla/c_src/exla/exla.cc
@@ -4,29 +4,19 @@
 #include "exla_nif_util.h"
 #include "exla_client.h"
 #include "exla_log_sink.h"
+#include "exla_ops.h"
 #include "xla/service/platform_util.h"
 #include "xla/shape_util.h"
 #include "xla/client/xla_builder.h"
 #include "xla/client/xla_computation.h"
 #include "xla/client/client.h"
-#include "xla/client/lib/math.h"
-#include "xla/client/lib/lu_decomposition.h"
-#include "xla/client/lib/qr.h"
-#include "xla/client/lib/self_adjoint_eig.h"
-#include "xla/client/lib/svd.h"
-#include "xla/client/lib/sorting.h"
 #include "xla/primitive_util.h"
 #include "xla/pjrt/pjrt_api.h"
-#include "xla/translate/hlo_to_mhlo/hlo_to_mlir_hlo.h"
-#include "mlir/IR/MLIRContext.h" 
-#include "xla/mlir_hlo/mhlo/IR/hlo_ops.h"
-#include "mlir/Dialect/Func/IR/FuncOps.h"  // from @llvm-project
-#include "llvm/Support/raw_ostream.h"
 
-struct ModuleResource {
-  mlir::MLIRContext context;
-  mlir::OwningOpRef<mlir::ModuleOp> module;
-};
+// MLIR
+#include <llvm/Support/raw_os_ostream.h>
+#include <iostream>
+#include "mlir/builder.h"
 
 // All of these are created with calls to `new` and subsequently
 // passed to the VM as pointers-to-pointers so we balance it out
@@ -76,9 +66,6 @@ static int open_resources(ErlNifEnv* env) {
   if (!exla::nif::open_resource<xla::XlaComputation>(env, mod, "Computation")) {
     return -1;
   }
-  if (!exla::nif::open_resource<struct ModuleResource>(env, mod, "ModuleOp")) {
-    return -1;
-  }
   if (!exla::nif::open_resource<exla::ExlaExecutable*>(env, mod, "Executable", free_exla_executable)) {
     return -1;
   }
@@ -91,6 +78,16 @@ static int open_resources(ErlNifEnv* env) {
   if (!exla::nif::open_resource<exla::ExlaBuffer*>(env, mod, "ExlaBuffer", free_exla_buffer)) {
     return -1;
   }
+  if (!exla::nif::open_resource<exla::MLIRFunction*>(env, mod, "MLIRBlock")) {
+    return -1;
+  }
+  if (!exla::nif::open_resource<mlir::Value>(env, mod, "MLIRValue")) {
+    return -1;
+  }
+  // MLIR
+  if (!exla::nif::open_resource<exla::MLIRModule *>(env, mod, "ExlaMLIRModule")) {
+    return -1;
+  }
   return 1;
 }
 
@@ -99,6 +96,194 @@ static int load(ErlNifEnv* env, void** priv, ERL_NIF_TERM load_info) {
 
   return 0;
 }
+
+// MLIR Builder Functions
+ERL_NIF_TERM new_mlir_module(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
+  if (argc != 0) {
+    return exla::nif::error(env, "Bad argument count.");
+  }
+
+  exla::MLIRModule * module = new exla::MLIRModule();
+
+  return exla::nif::ok(env, exla::nif::make<exla::MLIRModule *>(env, module));
+}
+
+ERL_NIF_TERM create_mlir_function(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
+  if (argc != 4) {
+    return exla::nif::error(env, "Bad argument count.");
+  }
+
+  exla::MLIRModule ** module;
+  std::string func_name;
+  std::vector<std::pair<std::vector<exla::int64>, int>> arg_types;
+  std::pair<std::vector<exla::int64>, int> ret_type;
+
+  if (!exla::nif::get<exla::MLIRModule*>(env, argv[0], module)) {
+    return exla::nif::error(env, "Unable to get module.");
+  }
+  if (!exla::nif::get(env, argv[1], func_name)) {
+    return exla::nif::error(env, "Unable to get function name.");
+  }
+
+  unsigned num_args;
+  if (!enif_get_list_length(env, argv[2], &num_args)) {
+    return exla::nif::error(env, "Unable to get args.");
+  }
+  arg_types.reserve(num_args);
+
+  ERL_NIF_TERM cur, head, tail;
+  cur = argv[2];
+  while (enif_get_list_cell(env, cur, &head, &tail)) {
+    int arity;
+    const ERL_NIF_TERM * tuple;
+    if (!enif_get_tuple(env, head, &arity, &tuple) || arity != 2) {
+      return exla::nif::error(env, "Unable to get args.");
+    }
+
+    std::vector<exla::int64> dims;
+    int type_int;
+
+    if (!exla::nif::get_list(env, tuple[0], dims)) {
+      return exla::nif::error(env, "Unable to get args.");
+    }
+    if (!exla::nif::get(env, tuple[1], &type_int)) {
+      return exla::nif::error(env, "Unable to get type int.");
+    }
+
+    arg_types.emplace_back(dims, type_int);
+    cur = tail;
+  }
+
+  int ret_arity;
+  const ERL_NIF_TERM * ret_tuple;
+  if (!enif_get_tuple(env, head, &ret_arity, &ret_tuple) || ret_arity != 2) {
+    return exla::nif::error(env, "Unable to get return.");
+  }
+
+  std::vector<exla::int64> ret_dims;
+  int ret_type_int;
+
+  if (!exla::nif::get_list(env, ret_tuple[0], ret_dims)) {
+    return exla::nif::error(env, "Unable to get return.");
+  }
+  if (!exla::nif::get(env, ret_tuple[1], &ret_type_int)) {
+    return exla::nif::error(env, "Unable to get return.");
+  }
+
+  ret_type = std::make_pair(ret_dims, ret_type_int);
+
+  exla::MLIRFunction * func = (*module)->CreateFunction(func_name, arg_types, ret_type);
+
+  return exla::nif::ok(env, exla::nif::make<exla::MLIRFunction*>(env, func));
+}
+
+ERL_NIF_TERM get_mlir_function_arguments(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
+  if (argc != 1) {
+    return exla::nif::error(env, "Bad argument count.");
+  }
+
+  exla::MLIRFunction ** function;
+
+  if (!exla::nif::get<exla::MLIRFunction*>(env, argv[0], function)) {
+    return exla::nif::error(env, "Unable to get function.");
+  }
+
+  llvm::MutableArrayRef<mlir::BlockArgument> args = (*function)->get_arguments();
+  std::vector<ERL_NIF_TERM> terms;
+  terms.reserve(args.size());
+
+  for (auto arg : args) {
+    ERL_NIF_TERM term = exla::nif::make<mlir::Value>(env, arg);
+    terms.push_back(term);
+  }
+
+  return exla::nif::ok(env, enif_make_list_from_array(env, terms.data(), terms.size()));
+}
+
+ERL_NIF_TERM mlir_add(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
+  if (argc != 3) {
+    return exla::nif::error(env, "Bad argument count.");
+  }
+
+  exla::MLIRFunction ** function;
+  mlir::Value * lhs;
+  mlir::Value * rhs;
+
+  if (!exla::nif::get<exla::MLIRFunction*>(env, argv[0], function)) {
+    return exla::nif::error(env, "Unable to get function.");
+  }
+  if (!exla::nif::get<mlir::Value>(env, argv[1], lhs)) {
+    return exla::nif::error(env, "Unable to get lhs.");
+  }
+  if (!exla::nif::get<mlir::Value>(env, argv[2], rhs)) {
+    return exla::nif::error(env, "Unable to get rhs.");
+  }
+
+  mlir::Value res = (*function)->AddOp(*lhs, *rhs);
+
+  return exla::nif::ok(env, exla::nif::make<mlir::Value>(env, res));
+}
+
+ERL_NIF_TERM mlir_subtract(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
+  if (argc != 3) {
+    return exla::nif::error(env, "Bad argument count.");
+  }
+
+  exla::MLIRFunction ** function;
+  mlir::Value * lhs;
+  mlir::Value * rhs;
+
+  if (!exla::nif::get<exla::MLIRFunction*>(env, argv[0], function)) {
+    return exla::nif::error(env, "Unable to get function.");
+  }
+  if (!exla::nif::get<mlir::Value>(env, argv[1], lhs)) {
+    return exla::nif::error(env, "Unable to get lhs.");
+  }
+  if (!exla::nif::get<mlir::Value>(env, argv[2], rhs)) {
+    return exla::nif::error(env, "Unable to get rhs.");
+  }
+
+  mlir::Value res = (*function)->SubtractOp(*lhs, *rhs);
+
+  return exla::nif::ok(env, exla::nif::make<mlir::Value>(env, res));
+}
+
+ERL_NIF_TERM mlir_build(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
+  if (argc != 2) {
+    return exla::nif::error(env, "Bad argument count.");
+  }
+
+  exla::MLIRFunction ** function;
+  mlir::Value * root;
+
+  if (!exla::nif::get<exla::MLIRFunction*>(env, argv[0], function)) {
+    return exla::nif::error(env, "Unable to get function.");
+  }
+  if (!exla::nif::get<mlir::Value>(env, argv[1], root)) {
+    return exla::nif::error(env, "Unable to get root.");
+  }
+
+  (*function)->Build(*root);
+
+  return exla::nif::ok(env);
+}
+
+ERL_NIF_TERM dump_mlir_module(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
+  if (argc != 1) {
+    return exla::nif::error(env, "Bad argument count.");
+  }
+
+  exla::MLIRModule** builder;
+
+  if (!exla::nif::get<exla::MLIRModule*>(env, argv[0], builder)) {
+    return exla::nif::error(env, "Unable to get builder.");
+  }
+
+  (*builder)->module().dump();
+
+  return exla::nif::ok(env);
+}
+
 
 // XlaBuilder Functions
 
@@ -156,34 +341,6 @@ ERL_NIF_TERM build(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
     (*builder)->Build(*root), env);
 
   return exla::nif::ok(env, exla::nif::make<xla::XlaComputation>(env, computation));
-}
-
-ERL_NIF_TERM parameter(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  if (argc != 4) {
-    return exla::nif::error(env, "Bad argument count.");
-  }
-
-  xla::XlaBuilder** builder;
-  exla::int64 param_num;
-  xla::Shape* shape;
-  std::string name;
-
-  if (!exla::nif::get<xla::XlaBuilder*>(env, argv[0], builder)) {
-    return exla::nif::error(env, "Unable to get builder.");
-  }
-  if (!exla::nif::get(env, argv[1], &param_num)) {
-    return exla::nif::error(env, "Unable to get parameter number.");
-  }
-  if (!exla::nif::get<xla::Shape>(env, argv[2], shape)) {
-    return exla::nif::error(env, "Unable to get parameter shape.");
-  }
-  if (!exla::nif::get(env, argv[3], name)) {
-    return exla::nif::error(env, "Unable to get parameter name.");
-  }
-
-  xla::XlaOp op = xla::Parameter((*builder), param_num, *shape, name);
-
-  return exla::nif::ok(env, exla::nif::make<xla::XlaOp>(env, op));
 }
 
 // ExlaBuffer Functions
@@ -315,1735 +472,6 @@ ERL_NIF_TERM get_shape_info(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[])
   }
 
   return exla::nif::ok(env, exla::nif::make_shape_info(env, *shape));
-}
-
-// XlaOp Functions
-
-// Tuples
-
-ERL_NIF_TERM tuple(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  if (argc != 2) {
-    return exla::nif::error(env, "Bad argument count.");
-  }
-
-  xla::XlaBuilder** builder;
-  std::vector<xla::XlaOp> elements;
-
-  if (!exla::nif::get<xla::XlaBuilder*>(env, argv[0], builder)) {
-    return exla::nif::error(env, "Unable to get builder.");
-  }
-  if (!exla::nif::get_list<xla::XlaOp>(env, argv[1], elements)) {
-    return exla::nif::error(env, "Unable to get tuple elements.");
-  }
-
-  xla::XlaOp op = xla::Tuple(*builder, elements);
-
-  return exla::nif::ok(env, exla::nif::make<xla::XlaOp>(env, op));
-}
-
-ERL_NIF_TERM get_tuple_element(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  if (argc != 2) {
-    return exla::nif::error(env, "Bad argument count.");
-  }
-
-  xla::XlaOp* operand;
-  exla::int64 index;
-
-  if (!exla::nif::get<xla::XlaOp>(env, argv[0], operand)) {
-    return exla::nif::error(env, "Unable to get operand.");
-  }
-  if (!exla::nif::get(env, argv[1], &index)) {
-    return exla::nif::error(env, "Unable to get index.");
-  }
-
-  xla::XlaOp op = xla::GetTupleElement(*operand, index);
-
-  return exla::nif::ok(env, exla::nif::make<xla::XlaOp>(env, op));
-}
-
-// Conditionals
-
-ERL_NIF_TERM conditional_if(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  if (argc != 5) {
-    return exla::nif::error(env, "Bad argument count.");
-  }
-
-  xla::XlaOp* pred;
-  xla::XlaOp* true_op;
-  xla::XlaOp* false_op;
-  xla::XlaComputation* true_comp;
-  xla::XlaComputation* false_comp;
-
-  if (!exla::nif::get<xla::XlaOp>(env, argv[0], pred)) {
-    return exla::nif::error(env, "Unable to get predicate.");
-  }
-  if (!exla::nif::get<xla::XlaOp>(env, argv[1], true_op)) {
-    return exla::nif::error(env, "Unable to get true operand.");
-  }
-  if (!exla::nif::get<xla::XlaComputation>(env, argv[2], true_comp)) {
-    return exla::nif::error(env, "Unable to get true computation.");
-  }
-  if (!exla::nif::get<xla::XlaOp>(env, argv[3], false_op)) {
-    return exla::nif::error(env, "Unable to get false operand.");
-  }
-  if (!exla::nif::get<xla::XlaComputation>(env, argv[4], false_comp)) {
-    return exla::nif::error(env, "Unable to get false computation.");
-  }
-
-  xla::XlaOp op = xla::Conditional(*pred, *true_op,
-                                   *true_comp, *false_op,
-                                   *false_comp);
-
-  return exla::nif::ok(env, exla::nif::make<xla::XlaOp>(env, op));
-}
-
-ERL_NIF_TERM conditional_multi(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  if (argc != 3) {
-    return exla::nif::error(env, "Bad argument count.");
-  }
-
-  xla::XlaOp* index;
-  std::vector<xla::XlaComputation*> branches;
-  std::vector<xla::XlaOp> operands;
-
-  if (!exla::nif::get<xla::XlaOp>(env, argv[0], index)) {
-    return exla::nif::error(env, "Unable to get index.");
-  }
-  if (!exla::nif::get_list<xla::XlaComputation*>(env, argv[1], branches)) {
-    return exla::nif::error(env, "Unable to get branches.");
-  }
-  if (!exla::nif::get_list<xla::XlaOp>(env, argv[2], operands)) {
-    return exla::nif::error(env, "Unable to get operands.");
-  }
-
-  xla::XlaOp op = xla::Conditional(*index, branches, operands);
-
-  return exla::nif::ok(env, exla::nif::make<xla::XlaOp>(env, op));
-}
-
-ERL_NIF_TERM select(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  if (argc != 3) {
-    return exla::nif::error(env, "Bad argument count.");
-  }
-
-  xla::XlaOp* pred;
-  xla::XlaOp* on_true;
-  xla::XlaOp* on_false;
-
-  if (!exla::nif::get<xla::XlaOp>(env, argv[0], pred)) {
-    return exla::nif::error(env, "Unable to get predicate.");
-  }
-  if (!exla::nif::get<xla::XlaOp>(env, argv[1], on_true)) {
-    return exla::nif::error(env, "Unable to get predicate.");
-  }
-  if (!exla::nif::get<xla::XlaOp>(env, argv[2], on_false)) {
-    return exla::nif::error(env, "Unable to get predicate.");
-  }
-
-  xla::XlaOp op = xla::Select(*pred, *on_true, *on_false);
-
-  return exla::nif::ok(env, exla::nif::make<xla::XlaOp>(env, op));
-}
-
-// Slicing
-
-ERL_NIF_TERM slice(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  if (argc != 4) {
-    return exla::nif::error(env, "Bad argument count.");
-  }
-
-  xla::XlaOp* operand;
-  std::vector<exla::int64> start_indices;
-  std::vector<exla::int64> limit_indices;
-  std::vector<exla::int64> strides;
-
-  if (!exla::nif::get<xla::XlaOp>(env, argv[0], operand)) {
-    return exla::nif::error(env, "Unable to get operand.");
-  }
-  if (!exla::nif::get_list(env, argv[1], start_indices)) {
-    return exla::nif::error(env, "Unable to get start indices.");
-  }
-  if (!exla::nif::get_list(env, argv[2], limit_indices)) {
-    return exla::nif::error(env, "Unable to get limit indices.");
-  }
-  if (!exla::nif::get_list(env, argv[3], strides)) {
-    return exla::nif::error(env, "Unable to get strides.");
-  }
-
-  xla::XlaOp op = xla::Slice(*operand, start_indices, limit_indices, strides);
-
-  return exla::nif::ok(env, exla::nif::make<xla::XlaOp>(env, op));
-}
-
-ERL_NIF_TERM dynamic_slice(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  if (argc != 3) {
-    return exla::nif::error(env, "Bad argument count.");
-  }
-
-  xla::XlaOp* operand;
-  std::vector<xla::XlaOp> start_indices;
-  std::vector<exla::int64> sizes;
-
-  if (!exla::nif::get(env, argv[0], operand)) {
-    return exla::nif::error(env, "Unable to get operand.");
-  }
-  if (!exla::nif::get_list<xla::XlaOp>(env, argv[1], start_indices)) {
-    return exla::nif::error(env, "Unable to get start index ops.");
-  }
-  if (!exla::nif::get_list(env, argv[2], sizes)) {
-    return exla::nif::error(env, "Unable to get sizes.");
-  }
-
-  xla::XlaOp op = xla::DynamicSlice(*operand, start_indices, sizes);
-
-  return exla::nif::ok(env, exla::nif::make<xla::XlaOp>(env, op));
-}
-
-ERL_NIF_TERM dynamic_update_slice(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  if (argc != 3) {
-    return exla::nif::error(env, "Bad argument count.");
-  }
-
-  xla::XlaOp* operand;
-  xla::XlaOp* update;
-  std::vector<xla::XlaOp> start_indices;
-
-  if (!exla::nif::get<xla::XlaOp>(env, argv[0], operand)) {
-    return exla::nif::error(env, "Unable to get operand.");
-  }
-  if (!exla::nif::get<xla::XlaOp>(env, argv[1], update)) {
-    return exla::nif::error(env, "Unable to get update.");
-  }
-  if (!exla::nif::get_list<xla::XlaOp>(env, argv[2], start_indices)) {
-    return exla::nif::error(env, "Unable to get start indices.");
-  }
-
-  xla::XlaOp op = xla::DynamicUpdateSlice(*operand, *update, start_indices);
-
-  return exla::nif::ok(env, exla::nif::make<xla::XlaOp>(env, op));
-}
-
-ERL_NIF_TERM gather(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  if (argc != 7) {
-    return exla::nif::error(env, "Bad argument count.");
-  }
-
-  xla::XlaOp* operand;
-  xla::XlaOp* start_indices;
-  exla::int64 index_vector_dim;
-  std::vector<exla::int64> slice_sizes;
-  std::vector<exla::int64> offset_dims;
-  std::vector<exla::int64> collapsed_slice_dims;
-  std::vector<exla::int64> start_index_map;
-
-  if (!exla::nif::get<xla::XlaOp>(env, argv[0], operand)) {
-    return exla::nif::error(env, "Unable to get operand.");
-  }
-  if (!exla::nif::get<xla::XlaOp>(env, argv[1], start_indices)) {
-    return exla::nif::error(env, "Unable to get start indices.");
-  }
-  if (!exla::nif::get(env, argv[2], &index_vector_dim)) {
-    return exla::nif::error(env, "Unable to get index vector dim.");
-  }
-  if (!exla::nif::get_list(env, argv[3], slice_sizes)) {
-    return exla::nif::error(env, "Unable to get slice sizes.");
-  }
-  if (!exla::nif::get_list(env, argv[4], offset_dims)) {
-    return exla::nif::error(env, "Unable to get offset dims.");
-  }
-  if (!exla::nif::get_list(env, argv[5], collapsed_slice_dims)) {
-    return exla::nif::error(env, "Unable to get collapsed slice dims.");
-  }
-  if (!exla::nif::get_list(env, argv[6], start_index_map)) {
-    return exla::nif::error(env, "Unable to get start index map.");
-  }
-
-  xla::GatherDimensionNumbers dim_numbers;
-
-  dim_numbers.set_index_vector_dim(index_vector_dim);
-
-  for (auto dim : offset_dims) {
-    dim_numbers.add_offset_dims(dim);
-  }
-  for (auto dim : collapsed_slice_dims) {
-    dim_numbers.add_collapsed_slice_dims(dim);
-  }
-  for (auto dim : start_index_map) {
-    dim_numbers.add_start_index_map(dim);
-  }
-
-  xla::XlaOp op = xla::Gather(*operand, *start_indices, dim_numbers, slice_sizes);
-
-  return exla::nif::ok(env, exla::nif::make<xla::XlaOp>(env, op));
-}
-
-// Creation
-
-ERL_NIF_TERM rng_normal(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  if (argc != 3) {
-    return exla::nif::error(env, "Bad argument count.");
-  }
-
-  xla::XlaOp* mu;
-  xla::XlaOp* sigma;
-  xla::Shape* shape;
-
-  if (!exla::nif::get<xla::XlaOp>(env, argv[0], mu)) {
-    return exla::nif::error(env, "Unable to get mu.");
-  }
-  if (!exla::nif::get<xla::XlaOp>(env, argv[1], sigma)) {
-    return exla::nif::error(env, "Unable to get sigma.");
-  }
-  if (!exla::nif::get<xla::Shape>(env, argv[2], shape)) {
-    return exla::nif::error(env, "Unable to get shape.");
-  }
-
-  xla::XlaOp op = xla::RngNormal(*mu, *sigma, *shape);
-
-  return exla::nif::ok(env, exla::nif::make<xla::XlaOp>(env, op));
-}
-
-ERL_NIF_TERM rng_uniform(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  if (argc != 3) {
-    return exla::nif::error(env, "Bad argument count.");
-  }
-
-  xla::XlaOp* a;
-  xla::XlaOp* b;
-  xla::Shape* shape;
-
-  if (!exla::nif::get<xla::XlaOp>(env, argv[0], a)) {
-    return exla::nif::error(env, "Unable to get mu.");
-  }
-  if (!exla::nif::get<xla::XlaOp>(env, argv[1], b)) {
-    return exla::nif::error(env, "Unable to get sigma.");
-  }
-  if (!exla::nif::get<xla::Shape>(env, argv[2], shape)) {
-    return exla::nif::error(env, "Unable to get shape.");
-  }
-
-  xla::XlaOp op = xla::RngUniform(*a, *b, *shape);
-
-  return exla::nif::ok(env, exla::nif::make<xla::XlaOp>(env, op));
-}
-
-ERL_NIF_TERM iota(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  if (argc != 3) {
-    return exla::nif::error(env, "Bad argument count.");
-  }
-
-  xla::XlaBuilder** builder;
-  xla::Shape* shape;
-  exla::int64 dimension;
-
-  if (!exla::nif::get<xla::XlaBuilder*>(env, argv[0], builder)) {
-    return exla::nif::error(env, "Unable to get builder.");
-  }
-  if (!exla::nif::get<xla::Shape>(env, argv[1], shape)) {
-    return exla::nif::error(env, "Unable to get shape.");
-  }
-  if (!exla::nif::get(env, argv[2], &dimension)) {
-    return exla::nif::error(env, "Unable to get dimension");
-  }
-
-  xla::XlaOp op = xla::Iota(*builder, *shape, dimension);
-
-  return exla::nif::ok(env, exla::nif::make<xla::XlaOp>(env, op));
-}
-
-// Binary Ops
-
-ERL_NIF_TERM xla_binary_op(ErlNifEnv* env,
-                           int argc,
-                           const ERL_NIF_TERM argv[],
-                           xla::XlaOp(*lambda)(xla::XlaOp, xla::XlaOp, absl::Span<const exla::int64>)) {
-  if (argc != 3) {
-    return exla::nif::error(env, "Bad argument count.");
-  }
-
-  xla::XlaOp *lhs;
-  xla::XlaOp *rhs;
-  std::vector<exla::int64> broadcast_dims;
-
-  if (!exla::nif::get<xla::XlaOp>(env, argv[0], lhs)) {
-    return exla::nif::error(env, "Unable to get left-hand side.");
-  }
-  if (!exla::nif::get<xla::XlaOp>(env, argv[1], rhs)) {
-    return exla::nif::error(env, "Unable to get right-hand side.");
-  }
-  if (!exla::nif::get_tuple(env, argv[2], broadcast_dims)) {
-    return exla::nif::error(env, "Unable to get broadcast dimensions.");
-  }
-
-  xla::XlaOp op = lambda(*lhs, *rhs, broadcast_dims);
-
-  return exla::nif::ok(env, exla::nif::make<xla::XlaOp>(env, op));
-}
-
-ERL_NIF_TERM add(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  return xla_binary_op(env, argc, argv, xla::Add);
-}
-
-ERL_NIF_TERM sub(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  return xla_binary_op(env, argc, argv, xla::Sub);
-}
-
-ERL_NIF_TERM mul(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  return xla_binary_op(env, argc, argv, xla::Mul);
-}
-
-ERL_NIF_TERM div(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  return xla_binary_op(env, argc, argv, xla::Div);
-}
-
-ERL_NIF_TERM rem(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  return xla_binary_op(env, argc, argv, xla::Rem);
-}
-
-ERL_NIF_TERM min(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  return xla_binary_op(env, argc, argv, xla::Min);
-}
-
-ERL_NIF_TERM max(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  return xla_binary_op(env, argc, argv, xla::Max);
-}
-
-ERL_NIF_TERM bitwise_and(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  return xla_binary_op(env, argc, argv, xla::And);
-}
-
-ERL_NIF_TERM bitwise_or(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  return xla_binary_op(env, argc, argv, xla::Or);
-}
-
-ERL_NIF_TERM bitwise_xor(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  return xla_binary_op(env, argc, argv, xla::Xor);
-}
-
-ERL_NIF_TERM shift_left(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  return xla_binary_op(env, argc, argv, xla::ShiftLeft);
-}
-
-ERL_NIF_TERM shift_right_logical(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  return xla_binary_op(env, argc, argv, xla::ShiftRightLogical);
-}
-
-ERL_NIF_TERM shift_right_arithmetic(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  return xla_binary_op(env, argc, argv, xla::ShiftRightArithmetic);
-}
-
-ERL_NIF_TERM equal(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  return xla_binary_op(env, argc, argv, xla::Eq);
-}
-
-ERL_NIF_TERM eq_total_order(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  return xla_binary_op(env, argc, argv, xla::EqTotalOrder);
-}
-
-ERL_NIF_TERM not_equal(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  return xla_binary_op(env, argc, argv, xla::Ne);
-}
-
-ERL_NIF_TERM ne_total_order(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  return xla_binary_op(env, argc, argv, xla::NeTotalOrder);
-}
-
-ERL_NIF_TERM greater_equal(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  return xla_binary_op(env, argc, argv, xla::Ge);
-}
-
-ERL_NIF_TERM greater(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  return xla_binary_op(env, argc, argv, xla::Gt);
-}
-
-ERL_NIF_TERM less(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  return xla_binary_op(env, argc, argv, xla::Lt);
-}
-
-ERL_NIF_TERM less_equal(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  return xla_binary_op(env, argc, argv, xla::Le);
-}
-
-ERL_NIF_TERM pow(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  return xla_binary_op(env, argc, argv, xla::Pow);
-}
-
-ERL_NIF_TERM complex(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  return xla_binary_op(env, argc, argv, xla::Complex);
-}
-
-ERL_NIF_TERM atan2(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  return xla_binary_op(env, argc, argv, xla::Atan2);
-}
-
-// Unary Ops
-
-ERL_NIF_TERM xla_unary_op(ErlNifEnv* env,
-                          int argc,
-                          const ERL_NIF_TERM argv[],
-                          xla::XlaOp(*lambda)(xla::XlaOp)) {
-  if (argc != 1) {
-    return exla::nif::error(env, "Bad argument count.");
-  }
-
-  xla::XlaOp *operand;
-
-  if (!exla::nif::get<xla::XlaOp>(env, argv[0], operand)) {
-    return exla::nif::error(env, "Unable to get operand.");
-  }
-
-  xla::XlaOp op = lambda(*operand);
-
-  return exla::nif::ok(env, exla::nif::make<xla::XlaOp>(env, op));
-}
-
-ERL_NIF_TERM abs(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  return xla_unary_op(env, argc, argv, xla::Abs);
-}
-
-ERL_NIF_TERM exp(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  return xla_unary_op(env, argc, argv, xla::Exp);
-}
-
-ERL_NIF_TERM expm1(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  return xla_unary_op(env, argc, argv, xla::Expm1);
-}
-
-ERL_NIF_TERM floor(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  return xla_unary_op(env, argc, argv, xla::Floor);
-}
-
-ERL_NIF_TERM ceil(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  return xla_unary_op(env, argc, argv, xla::Ceil);
-}
-
-ERL_NIF_TERM round(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  return xla_unary_op(env, argc, argv, xla::Round);
-}
-
-ERL_NIF_TERM log(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  return xla_unary_op(env, argc, argv, xla::Log);
-}
-
-ERL_NIF_TERM log1p(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  return xla_unary_op(env, argc, argv, xla::Log1p);
-}
-
-ERL_NIF_TERM sigmoid(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  return xla_unary_op(env, argc, argv, xla::Logistic);
-}
-
-ERL_NIF_TERM sign(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  return xla_unary_op(env, argc, argv, xla::Sign);
-}
-
-ERL_NIF_TERM clz(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  return xla_unary_op(env, argc, argv, xla::Clz);
-}
-
-ERL_NIF_TERM cos(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  return xla_unary_op(env, argc, argv, xla::Cos);
-}
-
-ERL_NIF_TERM sin(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  return xla_unary_op(env, argc, argv, xla::Sin);
-}
-
-ERL_NIF_TERM acos(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  return xla_unary_op(env, argc, argv, xla::Acos);
-}
-
-ERL_NIF_TERM asin(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  return xla_unary_op(env, argc, argv, xla::Asin);
-}
-
-ERL_NIF_TERM atan(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  return xla_unary_op(env, argc, argv, xla::Atan);
-}
-
-ERL_NIF_TERM cosh(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  return xla_unary_op(env, argc, argv, xla::Cosh);
-}
-
-ERL_NIF_TERM sinh(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  return xla_unary_op(env, argc, argv, xla::Sinh);
-}
-
-ERL_NIF_TERM tanh(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  return xla_unary_op(env, argc, argv, xla::Tanh);
-}
-
-ERL_NIF_TERM acosh(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  return xla_unary_op(env, argc, argv, xla::Acosh);
-}
-
-ERL_NIF_TERM asinh(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  return xla_unary_op(env, argc, argv, xla::Asinh);
-}
-
-ERL_NIF_TERM atanh(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  return xla_unary_op(env, argc, argv, xla::Atanh);
-}
-
-ERL_NIF_TERM real(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  return xla_unary_op(env, argc, argv, xla::Real);
-}
-
-ERL_NIF_TERM imag(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  return xla_unary_op(env, argc, argv, xla::Imag);
-}
-
-ERL_NIF_TERM sqrt(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  return xla_unary_op(env, argc, argv, xla::Sqrt);
-}
-
-ERL_NIF_TERM cbrt(ErlNifEnv *env, int argc, const ERL_NIF_TERM argv[])
-{
-  return xla_unary_op(env, argc, argv, xla::Cbrt);
-}
-
-ERL_NIF_TERM is_nan(ErlNifEnv *env, int argc, const ERL_NIF_TERM argv[])
-{
-  return xla_unary_op(env, argc, argv, xla::IsNan);
-}
-
-ERL_NIF_TERM is_infinity(ErlNifEnv *env, int argc, const ERL_NIF_TERM argv[])
-{
-  return xla_unary_op(env, argc, argv, [](xla::XlaOp op){return xla::IsInf(op);});
-}
-
-ERL_NIF_TERM execute_fft(ErlNifEnv *env, int argc, const ERL_NIF_TERM argv[], const xla::FftType fft_type)
-{
-  if (argc != 2)
-  {
-    return exla::nif::error(env, "Bad argument count.");
-  }
-
-  xla::XlaOp *operand;
-  exla::int64 fft_size;
-
-  if (!exla::nif::get<xla::XlaOp>(env, argv[0], operand))
-  {
-    return exla::nif::error(env, "Unable to get operand.");
-  }
-
-  if (!exla::nif::get(env, argv[1], &fft_size))
-  {
-    return exla::nif::error(env, "Unable to get fft_size.");
-  }
-
-  xla::XlaOp op = xla::Fft(*operand, fft_type, {fft_size});
-
-  return exla::nif::ok(env, exla::nif::make<xla::XlaOp>(env, op));
-}
-
-ERL_NIF_TERM fft(ErlNifEnv *env, int argc, const ERL_NIF_TERM argv[])
-{
-  return execute_fft(env, argc, argv, xla::FftType::FFT);
-}
-
-ERL_NIF_TERM ifft(ErlNifEnv *env, int argc, const ERL_NIF_TERM argv[])
-{
-  return execute_fft(env, argc, argv, xla::FftType::IFFT);
-}
-
-ERL_NIF_TERM rsqrt(ErlNifEnv *env, int argc, const ERL_NIF_TERM argv[])
-{
-  return xla_unary_op(env, argc, argv, xla::Rsqrt);
-}
-
-ERL_NIF_TERM erf(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  return xla_unary_op(env, argc, argv, xla::Erf);
-}
-
-ERL_NIF_TERM erfc(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  return xla_unary_op(env, argc, argv, xla::Erfc);
-}
-
-ERL_NIF_TERM erf_inv(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  return xla_unary_op(env, argc, argv, xla::ErfInv);
-}
-
-ERL_NIF_TERM is_finite(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  return xla_unary_op(env, argc, argv, xla::IsFinite);
-}
-
-ERL_NIF_TERM bitwise_not(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  return xla_unary_op(env, argc, argv, xla::Not);
-}
-
-ERL_NIF_TERM neg(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  return xla_unary_op(env, argc, argv, xla::Neg);
-}
-
-ERL_NIF_TERM conj(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  return xla_unary_op(env, argc, argv, xla::Conj);
-}
-
-ERL_NIF_TERM population_count(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  return xla_unary_op(env, argc, argv, xla::PopulationCount);
-}
-
-// Constants
-
-ERL_NIF_TERM constant_r0(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  if (argc != 3) {
-    return exla::nif::error(env, "Bad argument count.");
-  }
-
-  xla::XlaBuilder** builder;
-  xla::PrimitiveType type;
-
-  ERL_NIF_TERM term = argv[1];
-
-  if (!exla::nif::get<xla::XlaBuilder*>(env, argv[0], builder)) {
-    return exla::nif::error(env, "Unable to get builder.");
-  }
-  if (!exla::nif::get_primitive_type(env, argv[2], &type)) {
-    return exla::nif::error(env, "Unable to cast scalar to type.");
-  }
-
-  xla::XlaOp op;
-
-  switch (type) {
-    case xla::PrimitiveType::PRED:
-      op = xla::ConstantR0(*builder, exla::nif::get_value<xla::PrimitiveType::PRED>(env, term));
-      break;
-    case xla::PrimitiveType::U8:
-      op = xla::ConstantR0(*builder, exla::nif::get_value<xla::PrimitiveType::U8>(env, term));
-      break;
-    case xla::PrimitiveType::U16:
-      op = xla::ConstantR0(*builder, exla::nif::get_value<xla::PrimitiveType::U16>(env, term));
-      break;
-    case xla::PrimitiveType::U32:
-      op = xla::ConstantR0(*builder, exla::nif::get_value<xla::PrimitiveType::U32>(env, term));
-      break;
-    case xla::PrimitiveType::U64:
-      op = xla::ConstantR0(*builder, exla::nif::get_value<xla::PrimitiveType::U64>(env, term));
-      break;
-    case xla::PrimitiveType::S8:
-      op = xla::ConstantR0(*builder, exla::nif::get_value<xla::PrimitiveType::S8>(env, term));
-      break;
-    case xla::PrimitiveType::S16:
-      op = xla::ConstantR0(*builder, exla::nif::get_value<xla::PrimitiveType::S16>(env, term));
-      break;
-    case xla::PrimitiveType::S32:
-      op = xla::ConstantR0(*builder, exla::nif::get_value<xla::PrimitiveType::S32>(env, term));
-      break;
-    case xla::PrimitiveType::S64:
-      op = xla::ConstantR0(*builder, exla::nif::get_value<xla::PrimitiveType::S64>(env, term));
-      break;
-    case xla::PrimitiveType::F16:
-    {
-      // Not sure why the pattern breaks here
-      double var;
-      if (!enif_get_double(env, term, &var)) return exla::nif::error(env, "Unable to get constant");
-      exla::float16 val = static_cast<exla::float16>(var);
-      op = xla::ConstantR0(*builder, val);
-      break;
-    }
-    case xla::PrimitiveType::BF16:
-      op = xla::ConstantR0(*builder, exla::nif::get_value<xla::PrimitiveType::BF16>(env, term));
-      break;
-    case xla::PrimitiveType::F32:
-      op = xla::ConstantR0(*builder, exla::nif::get_value<xla::PrimitiveType::F32>(env, term));
-      break;
-    case xla::PrimitiveType::F64:
-      op = xla::ConstantR0(*builder, exla::nif::get_value<xla::PrimitiveType::F64>(env, term));
-      break;
-    case xla::PrimitiveType::C64:
-      return exla::nif::error(env, "Unsupported constant type.");
-    case xla::PrimitiveType::C128:
-      return exla::nif::error(env, "Unsupported constant type.");
-    default:
-      return exla::nif::error(env, "Invalid type.");
-  }
-
-  return exla::nif::ok(env, exla::nif::make<xla::XlaOp>(env, op));
-}
-
-ERL_NIF_TERM constant_from_binary(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  if (argc != 3) {
-    return exla::nif::error(env, "Bad argument count.");
-  }
-
-  xla::XlaBuilder** builder;
-  ErlNifBinary binary;
-  xla::Shape* shape;
-
-  if (!exla::nif::get<xla::XlaBuilder*>(env, argv[0], builder)) {
-    return exla::nif::error(env, "Unable to get builder.");
-  }
-  if (!exla::nif::get_binary(env, argv[1], &binary)) {
-    return exla::nif::error(env, "Unable to get data.");
-  }
-  if (!exla::nif::get<xla::Shape>(env, argv[2], shape)) {
-    return exla::nif::error(env, "Unable to get shape.");
-  }
-
-  char * data = const_cast<char*>(reinterpret_cast<char*>(binary.data));
-  xla::BorrowingLiteral literal(data, *shape);
-
-  xla::XlaOp op = xla::ConstantLiteral(*builder, literal);
-
-  return exla::nif::ok(env, exla::nif::make<xla::XlaOp>(env, op));
-}
-
-// Functional Ops
-
-ERL_NIF_TERM reduce(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  if (argc != 4) {
-    return exla::nif::error(env, "Bad argument count.");
-  }
-
-  xla::XlaOp* operand;
-  xla::XlaOp* init_value;
-  xla::XlaComputation* computation;
-  std::vector<exla::int64> dimensions_to_reduce;
-
-  if (!exla::nif::get<xla::XlaOp>(env, argv[0], operand)) {
-    return exla::nif::error(env, "Unable to get operand.");
-  }
-  if (!exla::nif::get<xla::XlaOp>(env, argv[1], init_value)) {
-    return exla::nif::error(env, "Unable to get initial value.");
-  }
-  if (!exla::nif::get<xla::XlaComputation>(env, argv[2], computation)) {
-    return exla::nif::error(env, "Unable to get computation.");
-  }
-  if (!exla::nif::get_tuple(env, argv[3], dimensions_to_reduce)) {
-    return exla::nif::error(env, "Unable to get reduction dimensions.");
-  }
-
-  xla::XlaOp op = xla::Reduce(*operand, *init_value,
-                              *computation, dimensions_to_reduce);
-
-  return exla::nif::ok(env, exla::nif::make<xla::XlaOp>(env, op));
-}
-
-ERL_NIF_TERM variadic_reduce(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  if (argc != 5) {
-    return exla::nif::error(env, "Bad argument count.");
-  }
-
-  xla::XlaBuilder** builder;
-  std::vector<xla::XlaOp> operands;
-  std::vector<xla::XlaOp> init_values;
-  xla::XlaComputation* computation;
-  std::vector<exla::int64> dimensions_to_reduce;
-
-  if (!exla::nif::get<xla::XlaBuilder*>(env, argv[0], builder)) {
-    return exla::nif::error(env, "Unable to get builder.");
-  }
-  if (!exla::nif::get_list<xla::XlaOp>(env, argv[1], operands)) {
-    return exla::nif::error(env, "Unable to get operands.");
-  }
-  if (!exla::nif::get_list<xla::XlaOp>(env, argv[2], init_values)) {
-    return exla::nif::error(env, "Unable to get initial values.");
-  }
-  if (!exla::nif::get<xla::XlaComputation>(env, argv[3], computation)) {
-    return exla::nif::error(env, "Unable to get computation.");
-  }
-  if (!exla::nif::get_tuple(env, argv[4], dimensions_to_reduce)) {
-    return exla::nif::error(env, "Unable to get dimensions.");
-  }
-
-  xla::XlaOp op = xla::Reduce(*builder, operands,
-                              init_values, *computation,
-                              dimensions_to_reduce);
-
-  return exla::nif::ok(env, exla::nif::make<xla::XlaOp>(env, op));
-}
-
-ERL_NIF_TERM window_reduce(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  if (argc != 7) {
-    return exla::nif::error(env, "Bad argument count.");
-  }
-
-  xla::XlaOp* operand;
-  xla::XlaOp* initial_value;
-  xla::XlaComputation* computation;
-  std::vector<exla::int64> window_dimensions;
-  std::vector<exla::int64> window_strides;
-  // std::vector<exla::int64> base_dilations;
-  std::vector<exla::int64> window_dilations;
-  std::vector<std::pair<exla::int64, exla::int64>> padding_config;
-
-  if (!exla::nif::get<xla::XlaOp>(env, argv[0], operand)) {
-    return exla::nif::error(env, "Unable to get operand.");
-  }
-  if (!exla::nif::get<xla::XlaOp>(env, argv[1], initial_value)) {
-    return exla::nif::error(env, "Unable to get initial value.");
-  }
-  if (!exla::nif::get<xla::XlaComputation>(env, argv[2], computation)) {
-    return exla::nif::error(env, "Unable to get computation.");
-  }
-  if (!exla::nif::get_tuple(env, argv[3], window_dimensions)) {
-    return exla::nif::error(env, "Unable to get window dimensions.");
-  }
-  if (!exla::nif::get_list(env, argv[4], window_strides)) {
-    return exla::nif::error(env, "Unable to get window strides.");
-  }
-  if (!exla::nif::get_list(env, argv[5], window_dilations)) {
-    return exla::nif::error(env, "Unable to get window dilations.");
-  }
-  if (!exla::nif::get_general_padding(env, argv[6], padding_config)) {
-    return exla::nif::error(env, "Unable to get padding configuration.");
-  }
-
-  xla::XlaOp op = xla::ReduceWindowWithGeneralPadding(*operand,
-                                                      *initial_value,
-                                                      *computation,
-                                                      window_dimensions,
-                                                      window_strides,
-                                                      {},
-                                                      window_dilations,
-                                                      padding_config);
-
-  return exla::nif::ok(env, exla::nif::make<xla::XlaOp>(env, op));
-}
-
-ERL_NIF_TERM select_and_scatter(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  if (argc != 8) {
-    return exla::nif::error(env, "Bad argument count.");
-  }
-
-  xla::XlaOp* operand;
-  xla::XlaComputation* select_fn;
-  std::vector<exla::int64> window_dimensions;
-  std::vector<exla::int64> window_strides;
-  std::vector<std::pair<exla::int64, exla::int64>> padding_config;
-  xla::XlaOp* source;
-  xla::XlaOp* init_value;
-  xla::XlaComputation* scatter_fn;
-
-  if (!exla::nif::get<xla::XlaOp>(env, argv[0], operand)) {
-    return exla::nif::error(env, "Unable to get operand.");
-  }
-  if (!exla::nif::get<xla::XlaComputation>(env, argv[1], select_fn)) {
-    return exla::nif::error(env, "Unable to get select function.");
-  }
-  if (!exla::nif::get_tuple(env, argv[2], window_dimensions)) {
-    return exla::nif::error(env, "Unable to get window dimensions.");
-  }
-  if (!exla::nif::get_list(env, argv[3], window_strides)) {
-    return exla::nif::error(env, "Unable to get window strides.");
-  }
-  if (!exla::nif::get_general_padding(env, argv[4], padding_config)) {
-    return exla::nif::error(env, "Unable to get padding configuration.");
-  }
-  if (!exla::nif::get<xla::XlaOp>(env, argv[5], source)) {
-    return exla::nif::error(env, "Unable to get source.");
-  }
-  if (!exla::nif::get<xla::XlaOp>(env, argv[6], init_value)) {
-    return exla::nif::error(env, "Unable to get initial value.");
-  }
-  if (!exla::nif::get<xla::XlaComputation>(env, argv[7], scatter_fn)) {
-    return exla::nif::error(env, "Unable to get scatter function.");
-  }
-
-  xla::XlaOp op = xla::SelectAndScatterWithGeneralPadding(*operand,
-                                                          *select_fn,
-                                                          window_dimensions,
-                                                          window_strides,
-                                                          padding_config,
-                                                          *source,
-                                                          *init_value,
-                                                          *scatter_fn);
-
-  return exla::nif::ok(env, exla::nif::make<xla::XlaOp>(env, op));
-}
-
-ERL_NIF_TERM scatter(ErlNifEnv *env, int argc, const ERL_NIF_TERM argv[])
-{
-  if (argc != 8)
-  {
-    return exla::nif::error(env, "Bad argument count.");
-  }
-
-  xla::XlaOp *target;
-  xla::XlaOp *indices;
-  xla::XlaOp *updates;
-  xla::XlaComputation *scatter_fn;
-  exla::int64 index_vector_dim;
-  std::vector<exla::int64> update_window_dims;
-  std::vector<exla::int64> inserted_window_dims;
-  std::vector<exla::int64> scatter_dims_to_operand_dims;
-
-  if (!exla::nif::get<xla::XlaOp>(env, argv[0], target))
-  {
-    return exla::nif::error(env, "Unable to get target.");
-  }
-  if (!exla::nif::get<xla::XlaOp>(env, argv[1], indices))
-  {
-    return exla::nif::error(env, "Unable to get indices.");
-  }
-  if (!exla::nif::get<xla::XlaOp>(env, argv[2], updates))
-  {
-    return exla::nif::error(env, "Unable to get updates.");
-  }
-  if (!exla::nif::get<xla::XlaComputation>(env, argv[3], scatter_fn))
-  {
-    return exla::nif::error(env, "Unable to get scatter function.");
-  }
-  if (!exla::nif::get(env, argv[4], &index_vector_dim))
-  {
-    return exla::nif::error(env, "Unable to get index vector dim.");
-  }
-  if (!exla::nif::get_list(env, argv[5], update_window_dims))
-  {
-    return exla::nif::error(env, "Unable to get update window dims.");
-  }
-  if (!exla::nif::get_list(env, argv[6], inserted_window_dims))
-  {
-    return exla::nif::error(env, "Unable to get update window dims.");
-  }
-  if (!exla::nif::get_list(env, argv[7], scatter_dims_to_operand_dims))
-  {
-    return exla::nif::error(env, "Unable to get update window dims.");
-  }
-
-  xla::ScatterDimensionNumbers scatter_dim_numbers;
-
-  scatter_dim_numbers.set_index_vector_dim(index_vector_dim);
-
-
-  for (size_t i = 0; i < update_window_dims.size(); i++)
-  {
-    // scatter_dim_numbers.set_update_window_dims(i, update_window_dims[i]);
-    scatter_dim_numbers.add_update_window_dims(update_window_dims[i]);
-  }
-
-  for (size_t i = 0; i < inserted_window_dims.size(); i++)
-  {
-    scatter_dim_numbers.add_inserted_window_dims(inserted_window_dims[i]);
-  }
-
-  for (size_t i = 0; i < scatter_dims_to_operand_dims.size(); i++)
-  {
-    scatter_dim_numbers.add_scatter_dims_to_operand_dims(scatter_dims_to_operand_dims[i]);
-  }
-
-  xla::XlaOp op = xla::Scatter(
-      *target,
-      *indices,
-      *updates,
-      *scatter_fn,
-      scatter_dim_numbers,
-      false);
-
-  return exla::nif::ok(env, exla::nif::make<xla::XlaOp>(env, op));
-}
-
-ERL_NIF_TERM map(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  if (argc != 4) {
-    return exla::nif::error(env, "Bad argument count.");
-  }
-
-  xla::XlaBuilder** builder;
-  xla::XlaOp* operand;
-  xla::XlaComputation* computation;
-  std::vector<exla::int64> dimensions;
-
-  if (!exla::nif::get<xla::XlaBuilder*>(env, argv[0], builder)) {
-    return exla::nif::error(env, "Unable to get builder.");
-  }
-  if (!exla::nif::get<xla::XlaOp>(env, argv[1], operand)) {
-    return exla::nif::error(env, "Unable to get operand.");
-  }
-  if (!exla::nif::get<xla::XlaComputation>(env, argv[2], computation)) {
-    return exla::nif::error(env, "Unable to get computation.");
-  }
-  if (!exla::nif::get_list(env, argv[3], dimensions)) {
-    return exla::nif::error(env, "Unable to get map dimensions.");
-  }
-
-  xla::XlaOp op = xla::Map(*builder, {*operand}, *computation, dimensions);
-
-  return exla::nif::ok(env, exla::nif::make<xla::XlaOp>(env, op));
-}
-
-ERL_NIF_TERM while_loop(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  if (argc != 3) {
-    return exla::nif::error(env, "Bad argument count.");
-  }
-
-  xla::XlaComputation* cond_fn;
-  xla::XlaComputation* body_fn;
-  xla::XlaOp* init_value;
-
-  if (!exla::nif::get<xla::XlaComputation>(env, argv[0], cond_fn)) {
-    return exla::nif::error(env, "Unable to get condition computation.");
-  }
-  if (!exla::nif::get<xla::XlaComputation>(env, argv[1], body_fn)) {
-    return exla::nif::error(env, "Unable to get body computation.");
-  }
-  if (!exla::nif::get<xla::XlaOp>(env, argv[2], init_value)) {
-    return exla::nif::error(env, "Unable to get initial value.");
-  }
-
-  xla::XlaOp op = xla::While(*cond_fn, *body_fn, *init_value);
-
-  return exla::nif::ok(env, exla::nif::make<xla::XlaOp>(env, op));
-}
-
-ERL_NIF_TERM call(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  if (argc != 3) {
-    return exla::nif::error(env, "Bad argument count.");
-  }
-
-  xla::XlaBuilder **builder;
-  xla::XlaComputation *fn;
-  // We will always wrap fn to expect a single tuple-arg
-  // which contains all args because we can't call xla::Call
-  // in a dynamically variadic way.
-  std::vector<xla::XlaOp> args;
-
-  if (!exla::nif::get<xla::XlaBuilder*>(env, argv[0], builder)) {
-    return exla::nif::error(env, "Unable to get call builder.");
-  }
-  if (!exla::nif::get_list<xla::XlaOp>(env, argv[1], args)) {
-    return exla::nif::error(env, "Unable to get call args.");
-  }
-  if (!exla::nif::get<xla::XlaComputation>(env, argv[2], fn)) {
-    return exla::nif::error(env, "Unable to get body computation.");
-  }
-
-  xla::XlaOp op = xla::Call(*builder, *fn, args);
-
-  return exla::nif::ok(env, exla::nif::make<xla::XlaOp>(env, op));
-}
-
-// Shape/Type Manipulation
-
-ERL_NIF_TERM reshape(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  if (argc != 2) {
-    return exla::nif::error(env, "Bad argument count.");
-  }
-
-  xla::XlaOp* operand;
-  std::vector<exla::int64> new_shape;
-
-  if (!exla::nif::get<xla::XlaOp>(env, argv[0], operand)) {
-    return exla::nif::error(env, "Unable to get operand.");
-  }
-  if (!exla::nif::get_tuple(env, argv[1], new_shape)) {
-    return exla::nif::error(env, "Unable to get dimensions.");
-  }
-
-  xla::XlaOp op = xla::Reshape(*operand, new_shape);
-
-  return exla::nif::ok(env, exla::nif::make<xla::XlaOp>(env, op));
-}
-
-ERL_NIF_TERM broadcast_in_dim(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  if (argc != 3) {
-    return exla::nif::error(env, "Bad argument count.");
-  }
-
-  xla::XlaOp* operand;
-  std::vector<exla::int64> new_shape;
-  std::vector<exla::int64> broadcast_dims;
-
-  if (!exla::nif::get<xla::XlaOp>(env, argv[0], operand)) {
-    return exla::nif::error(env, "Unable to get operand.");
-  }
-  if (!exla::nif::get_tuple(env, argv[1], new_shape)) {
-    return exla::nif::error(env, "Unable to get dimensions.");
-  }
-  if (!exla::nif::get_tuple(env, argv[2], broadcast_dims)) {
-    return exla::nif::error(env, "Unable to get broadcast dimensions.");
-  }
-
-  xla::XlaOp op = xla::BroadcastInDim(*operand, new_shape, broadcast_dims);
-  return exla::nif::ok(env, exla::nif::make<xla::XlaOp>(env, op));
-}
-
-ERL_NIF_TERM get_shape_op(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  if (argc != 2) {
-    return exla::nif::error(env, "Bad argument count.");
-  }
-
-  xla::XlaBuilder** builder;
-  xla::XlaOp* operand;
-
-  if (!exla::nif::get<xla::XlaBuilder*>(env, argv[0], builder)) {
-    return exla::nif::error(env, "Unable to get builder.");
-  }
-  if (!exla::nif::get<xla::XlaOp>(env, argv[1], operand)) {
-    return exla::nif::error(env, "Unable to get operand.");
-  }
-
-  EXLA_ASSIGN_OR_RETURN_NIF(xla::Shape shape,
-    (*builder)->GetShape(*operand), env);
-
-  return exla::nif::ok(env, exla::nif::make<xla::Shape>(env, shape));
-}
-
-ERL_NIF_TERM convert_element_type(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  if (argc != 2) {
-    return exla::nif::error(env, "Bad argument count.");
-  }
-
-  xla::XlaOp* operand;
-  xla::PrimitiveType type;
-
-  if (!exla::nif::get<xla::XlaOp>(env, argv[0], operand)) {
-    return exla::nif::error(env, "Unable to get operand.");
-  }
-  if (!exla::nif::get_primitive_type(env, argv[1], &type)) {
-    return exla::nif::error(env, "Unable to get type string.");
-  }
-
-  xla::XlaOp op = xla::ConvertElementType(*operand, type);
-
-  return exla::nif::ok(env, exla::nif::make<xla::XlaOp>(env, op));
-}
-
-ERL_NIF_TERM bitcast_convert_type(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  if (argc != 2) {
-    return exla::nif::error(env, "Bad argument count.");
-  }
-
-  xla::XlaOp* operand;
-  xla::PrimitiveType type;
-
-  if (!exla::nif::get<xla::XlaOp>(env, argv[0], operand)) {
-    return exla::nif::error(env, "Unable to get operand.");
-  }
-  if (!exla::nif::get_primitive_type(env, argv[1], &type)) {
-    return exla::nif::error(env, "Unable to get type string.");
-  }
-
-  xla::XlaOp op = xla::BitcastConvertType(*operand, type);
-
-  return exla::nif::ok(env, exla::nif::make<xla::XlaOp>(env, op));
-}
-
-ERL_NIF_TERM transpose(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  if (argc != 2) {
-    return exla::nif::error(env, "Bad argument count.");
-  }
-
-  xla::XlaOp* operand;
-  std::vector<exla::int64> permutation;
-
-  if (!exla::nif::get<xla::XlaOp>(env, argv[0], operand)) {
-    return exla::nif::error(env, "Unable to get operand.");
-  }
-  if (!exla::nif::get_tuple(env, argv[1], permutation)) {
-    return exla::nif::error(env, "Unable to get permutation.");
-  }
-
-  xla::XlaOp op = xla::Transpose(*operand, permutation);
-
-  return exla::nif::ok(env, exla::nif::make<xla::XlaOp>(env, op));
-}
-
-// Other Ops
-
-ERL_NIF_TERM dot(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  if (argc != 3) {
-    return exla::nif::error(env, "Bad argument count.");
-  }
-
-  xla::XlaOp* lhs;
-  xla::XlaOp* rhs;
-  xla::PrecisionConfig config;
-
-  if (!exla::nif::get<xla::XlaOp>(env, argv[0], lhs)) {
-    return exla::nif::error(env, "Unable to get left-hand side operand.");
-  }
-  if (!exla::nif::get<xla::XlaOp>(env, argv[1], rhs)) {
-    return exla::nif::error(env, "Unable to get right-hand side operand.");
-  }
-  if (!exla::nif::get_precision_config(env, argv[2], 2, &config)) {
-    return exla::nif::error(env, "Unable to get precision configuration.");
-  }
-
-  xla::XlaOp op = xla::Dot(*lhs, *rhs, &config);
-  return exla::nif::ok(env, exla::nif::make<xla::XlaOp>(env, op));
-}
-
-ERL_NIF_TERM dot_general(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  if (argc != 4) {
-    return exla::nif::error(env, "Bad argument count.");
-  }
-
-  xla::XlaOp* lhs;
-  xla::XlaOp* rhs;
-  xla::DotDimensionNumbers dnums;
-  xla::PrecisionConfig config;
-
-  if (!exla::nif::get<xla::XlaOp>(env, argv[0], lhs)) {
-    return exla::nif::error(env, "Unable to get left-hand side operand.");
-  }
-  if (!exla::nif::get<xla::XlaOp>(env, argv[1], rhs)) {
-    return exla::nif::error(env, "Unable to get right-hand side operand.");
-  }
-  if (!exla::nif::get_dot_dimension_numbers(env, argv[2], &dnums)) {
-    return exla::nif::error(env, "Unable to get contraction dimensions.");
-  }
-  if (!exla::nif::get_precision_config(env, argv[3], 2, &config)) {
-    return exla::nif::error(env, "Unable to get precision configuration.");
-  }
-
-  xla::XlaOp op = xla::DotGeneral(*lhs, *rhs, dnums, &config);
-
-  return exla::nif::ok(env, exla::nif::make<xla::XlaOp>(env, op));
-}
-
-ERL_NIF_TERM conv_general_dilated(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  if (argc != 10) {
-    return exla::nif::error(env, "Bad argument count.");
-  }
-
-  xla::XlaOp* operand;
-  xla::XlaOp* kernel;
-  std::vector<exla::int64> strides;
-  std::vector<std::pair<exla::int64, exla::int64>> padding;
-  std::vector<exla::int64> lhs_dilation;
-  std::vector<exla::int64> rhs_dilation;
-  xla::ConvolutionDimensionNumbers dimension_numbers;
-  exla::int64 feature_group_count;
-  exla::int64 batch_group_count;
-  xla::PrecisionConfig config;
-
-  if (!exla::nif::get<xla::XlaOp>(env, argv[0], operand)) {
-    return exla::nif::error(env, "Unable to get operand.");
-  }
-  if (!exla::nif::get<xla::XlaOp>(env, argv[1], kernel)) {
-    return exla::nif::error(env, "Unable to get kernel.");
-  }
-  if (!exla::nif::get_list(env, argv[2], strides)) {
-    return exla::nif::error(env, "Unable to get strides.");
-  }
-  if (!exla::nif::get_general_padding(env, argv[3], padding)) {
-    return exla::nif::error(env, "Unable to get padding.");
-  }
-  if (!exla::nif::get_list(env, argv[4], lhs_dilation)) {
-    return exla::nif::error(env, "Unable to get operand dilation.");
-  }
-  if (!exla::nif::get_list(env, argv[5], rhs_dilation)) {
-    return exla::nif::error(env, "Unable to get kernel dilation.");
-  }
-  if (!exla::nif::get_conv_dimension_numbers(env, argv[6], &dimension_numbers)) {
-    return exla::nif::error(env, "Unable to get conv dimension numbers.");
-  }
-  if (!exla::nif::get(env, argv[7], &feature_group_count)) {
-    return exla::nif::error(env, "Unable to get feature groups.");
-  }
-  if (!exla::nif::get(env, argv[8], &batch_group_count)) {
-    return exla::nif::error(env, "Unable to get batch groups.");
-  }
-  if (!exla::nif::get_precision_config(env, argv[9], 2, &config)) {
-    return exla::nif::error(env, "Unable to get precision config.");
-  }
-
-  xla::XlaOp op = xla::ConvGeneralDilated(*operand, *kernel, strides,
-                                          padding, lhs_dilation, rhs_dilation,
-                                          dimension_numbers, feature_group_count,
-                                          batch_group_count, &config);
-
-  return exla::nif::ok(env, exla::nif::make<xla::XlaOp>(env, op));
-}
-
-ERL_NIF_TERM pad(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  if (argc != 3) {
-    return exla::nif::error(env, "Bad argument count.");
-  }
-
-  xla::XlaOp* operand;
-  xla::XlaOp* pad_value;
-  xla::PaddingConfig padding_config;
-
-  if (!exla::nif::get<xla::XlaOp>(env, argv[0], operand)) {
-    return exla::nif::error(env, "Unable to get operand.");
-  }
-  if (!exla::nif::get<xla::XlaOp>(env, argv[1], pad_value)) {
-    return exla::nif::error(env, "Unable to get value.");
-  }
-  if (!exla::nif::get_padding_config(env, argv[2], &padding_config)) {
-    return exla::nif::error(env, "Unable to get padding configuration.");
-  }
-
-  xla::XlaOp op = xla::Pad(*operand, *pad_value, padding_config);
-
-  return exla::nif::ok(env, exla::nif::make<xla::XlaOp>(env, op));
-}
-
-ERL_NIF_TERM clamp(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  if (argc != 3) {
-    return exla::nif::error(env, "Bad argument count.");
-  }
-
-  xla::XlaOp* operand;
-  xla::XlaOp* min;
-  xla::XlaOp* max;
-
-  if (!exla::nif::get<xla::XlaOp>(env, argv[0], operand)) {
-    return exla::nif::error(env, "Unable to get operand.");
-  }
-  if (!exla::nif::get<xla::XlaOp>(env, argv[1], min)) {
-    return exla::nif::error(env, "Unable to get min value.");
-  }
-  if (!exla::nif::get<xla::XlaOp>(env, argv[2], max)) {
-    return exla::nif::error(env, "Unable to get max value");
-  }
-
-  xla::XlaOp op = xla::Clamp(*min, *operand, *max);
-
-  return exla::nif::ok(env, exla::nif::make<xla::XlaOp>(env, op));
-}
-
-ERL_NIF_TERM reverse(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  if (argc != 2) {
-    return exla::nif::error(env, "Bad argument count.");
-  }
-
-  xla::XlaOp* operand;
-  std::vector<exla::int64> dimensions;
-
-  if (!exla::nif::get<xla::XlaOp>(env, argv[0], operand)) {
-    return exla::nif::error(env, "Unable to get operand.");
-  }
-  if (!exla::nif::get_list(env, argv[1], dimensions)) {
-    return exla::nif::error(env, "Unable to get dimensions.");
-  }
-
-  xla::XlaOp op = xla::Rev(*operand, dimensions);
-
-  return exla::nif::ok(env, exla::nif::make<xla::XlaOp>(env, op));
-}
-
-ERL_NIF_TERM concatenate(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  if (argc != 3) {
-    return exla::nif::error(env, "Bad argument count.");
-  }
-
-  xla::XlaBuilder** builder;
-  std::vector<xla::XlaOp> operands;
-  exla::int64 dimension;
-
-  if (!exla::nif::get<xla::XlaBuilder*>(env, argv[0], builder)) {
-    return exla::nif::error(env, "Unable to get builder.");
-  }
-  if (!exla::nif::get_list<xla::XlaOp>(env, argv[1], operands)) {
-    return exla::nif::error(env, "Unable to get operands.");
-  }
-  if (!exla::nif::get(env, argv[2], &dimension)) {
-    return exla::nif::error(env, "Unable to get dimension.");
-  }
-
-  xla::XlaOp op = xla::ConcatInDim(*builder, operands, dimension);
-
-  return exla::nif::ok(env, exla::nif::make<xla::XlaOp>(env, op));
-}
-
-ERL_NIF_TERM sort(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  if (argc != 3) {
-    return exla::nif::error(env, "Bad argument count.");
-  }
-
-  xla::XlaOp* operand;
-  xla::XlaComputation* comparator;
-  exla::int64 dimension;
-
-  if (!exla::nif::get<xla::XlaOp>(env, argv[0], operand)) {
-    return exla::nif::error(env, "Unable to get operand.");
-  }
-  if (!exla::nif::get<xla::XlaComputation>(env, argv[1], comparator)) {
-    return exla::nif::error(env, "Unable to get comparator.");
-  }
-  if (!exla::nif::get(env, argv[2], &dimension)) {
-    return exla::nif::error(env, "Unable to get dimension.");
-  }
-
-  xla::XlaOp op = xla::Sort({*operand}, *comparator, dimension, true);
-
-  return exla::nif::ok(env, exla::nif::make<xla::XlaOp>(env, op));
-}
-
-ERL_NIF_TERM top_k(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  if (argc != 2) {
-    return exla::nif::error(env, "Bad argument count.");
-  }
-
-  xla::XlaOp* operand;
-  exla::int64 k;
-
-  if (!exla::nif::get<xla::XlaOp>(env, argv[0], operand)) {
-    return exla::nif::error(env, "Unable to get operand.");
-  }
-  if (!exla::nif::get(env, argv[1], &k)) {
-    return exla::nif::error(env, "Unable to get k.");
-  }
-
-  xla::XlaOp op = xla::TopK(*operand, k);
-
-  return exla::nif::ok(env, exla::nif::make<xla::XlaOp>(env, op));
-}
-
-ERL_NIF_TERM variadic_sort(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  if (argc != 3) {
-    return exla::nif::error(env, "Bad argument count.");
-  }
-
-  std::vector<xla::XlaOp> operands;
-  xla::XlaComputation* comparator;
-  exla::int64 dimension;
-
-  if (!exla::nif::get_list<xla::XlaOp>(env, argv[0], operands)) {
-    return exla::nif::error(env, "Unable to get operands.");
-  }
-  if (!exla::nif::get<xla::XlaComputation>(env, argv[1], comparator)) {
-    return exla::nif::error(env, "Unable to get comparator.");
-  }
-  if (!exla::nif::get(env, argv[2], &dimension)) {
-    return exla::nif::error(env, "Unable to get dimension.");
-  }
-
-  xla::XlaOp op = xla::Sort(operands, *comparator, dimension, true);
-
-  return exla::nif::ok(env, exla::nif::make<xla::XlaOp>(env, op));
-}
-
-// LinAlg Functions
-
-ERL_NIF_TERM cholesky(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  if (argc != 1) {
-    return exla::nif::error(env, "Bad argument count.");
-  }
-
-  xla::XlaOp* operand;
-
-  if (!exla::nif::get<xla::XlaOp>(env, argv[0], operand)) {
-    return exla::nif::error(env, "Unable to get operand.");
-  }
-
-  xla::XlaOp op = xla::Cholesky(*operand, true);
-
-  return exla::nif::ok(env, exla::nif::make<xla::XlaOp>(env, op));
-}
-
-ERL_NIF_TERM eigh(ErlNifEnv *env, int argc, const ERL_NIF_TERM argv[])
-{
-  if (argc != 4)
-  {
-    return exla::nif::error(env, "Bad argument count.");
-  }
-
-  xla::XlaOp *operand;
-  bool lower;
-  double eps;
-  int64_t max_iter;
-
-  if (!exla::nif::get<xla::XlaOp>(env, argv[0], operand))
-  {
-    return exla::nif::error(env, "Unable to get operand.");
-  }
-  if (!exla::nif::get(env, argv[1], &lower))
-  {
-    return exla::nif::error(env, "Unable to get lower flag.");
-  }
-  if (!exla::nif::get(env, argv[2], &eps))
-  {
-    return exla::nif::error(env, "Unable to get eps.");
-  }
-  if (!exla::nif::get(env, argv[3], &max_iter))
-  {
-    return exla::nif::error(env, "Unable to get max_iter.");
-  }
-
-  xla::SelfAdjointEigResult eigh_result = xla::SelfAdjointEig(*operand,
-                                                              lower,
-                                                              max_iter,
-                                                              eps);
-
-  ERL_NIF_TERM v = exla::nif::make<xla::XlaOp>(env, eigh_result.v);
-  ERL_NIF_TERM w = exla::nif::make<xla::XlaOp>(env, eigh_result.w);
-
-  return exla::nif::ok(env, enif_make_tuple2(env, v, w));
-}
-
-ERL_NIF_TERM lu(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  if (argc != 1) {
-    return exla::nif::error(env, "Bad argument count.");
-  }
-
-  xla::XlaOp* operand;
-
-  if (!exla::nif::get<xla::XlaOp>(env, argv[0], operand)) {
-    return exla::nif::error(env, "Unable to get operand.");
-  }
-
-  xla::LuDecompositionResult lu_result = xla::LuDecomposition(*operand);
-
-  ERL_NIF_TERM lu = exla::nif::make<xla::XlaOp>(env, lu_result.lu);
-  ERL_NIF_TERM pivots = exla::nif::make<xla::XlaOp>(env, lu_result.pivots);
-  ERL_NIF_TERM permutation = exla::nif::make<xla::XlaOp>(env, lu_result.permutation);
-
-  return exla::nif::ok(env, enif_make_tuple3(env, lu, pivots, permutation));
-}
-
-ERL_NIF_TERM qr(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  if (argc != 2) {
-    return exla::nif::error(env, "Bad argument count.");
-  }
-
-  xla::XlaOp* operand;
-  bool full_matrices;
-
-  if (!exla::nif::get<xla::XlaOp>(env, argv[0], operand)) {
-    return exla::nif::error(env, "Unable to get operand.");
-  }
-  if (!exla::nif::get(env, argv[1], &full_matrices)) {
-    return exla::nif::error(env, "Unable to get full matrices flag.");
-  }
-
-  xla::XlaOp q, r;
-
-  QrExplicit(*operand, full_matrices, q, r);
-
-  ERL_NIF_TERM q_term = exla::nif::make<xla::XlaOp>(env, q);
-  ERL_NIF_TERM r_term = exla::nif::make<xla::XlaOp>(env, r);
-
-  return exla::nif::ok(env, enif_make_tuple2(env, q_term, r_term));
-}
-
-ERL_NIF_TERM triangular_solve(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  if (argc != 6) {
-    return exla::nif::error(env, "Bad argument count.");
-  }
-
-  xla::XlaOp* a;
-  xla::XlaOp* b;
-  bool left_side;
-  bool lower;
-  bool unit_diagonal;
-  int transpose_a_int;
-
-  if (!exla::nif::get<xla::XlaOp>(env, argv[0], a)) {
-    return exla::nif::error(env, "Unable to get A.");
-  }
-  if (!exla::nif::get<xla::XlaOp>(env, argv[1], b)) {
-    return exla::nif::error(env, "Unable to get B.");
-  }
-  if (!exla::nif::get(env, argv[2], &left_side)) {
-    return exla::nif::error(env, "Unable to get left side flag.");
-  }
-  if (!exla::nif::get(env, argv[3], &lower)) {
-    return exla::nif::error(env, "Unable to get lower flag.");
-  }
-  if (!exla::nif::get(env, argv[4], &unit_diagonal)) {
-    return exla::nif::error(env, "Unable to get unit diagonal flag.");
-  }
-  if (!exla::nif::get(env, argv[5], &transpose_a_int)) {
-    return exla::nif::error(env, "Unable to get triangular solve transpose options.");
-  }
-
-  xla::TriangularSolveOptions::Transpose transpose_a;
-  switch (transpose_a_int) {
-    case 0:
-      transpose_a = xla::TriangularSolveOptions::NO_TRANSPOSE;
-      break;
-    case 1:
-      transpose_a = xla::TriangularSolveOptions::TRANSPOSE;
-      break;
-    case 2:
-      transpose_a = xla::TriangularSolveOptions::ADJOINT;
-      break;
-    default:
-      LOG(ERROR) << "Invalid triangular solve options.";
-      return 0;
-  }
-
-  xla::XlaOp op = xla::TriangularSolve(*a, *b,
-                                       left_side,
-                                       lower,
-                                       unit_diagonal,
-                                       transpose_a);
-
-  return exla::nif::ok(env, exla::nif::make<xla::XlaOp>(env, op));
-}
-
-// Infeed/Outfeed
-
-ERL_NIF_TERM infeed(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  if (argc != 2) {
-    return exla::nif::error(env, "Bad argument count.");
-  }
-
-  xla::XlaOp* token;
-  xla::Shape* shape;
-
-  if (!exla::nif::get<xla::XlaOp>(env, argv[0], token)) {
-    return exla::nif::error(env, "Unable to get token.");
-  }
-  if (!exla::nif::get<xla::Shape>(env, argv[1], shape)) {
-    return exla::nif::error(env, "Unable to get shape.");
-  }
-
-  // TODO(seanmor5): Determine if default config is optimal
-  xla::XlaOp op = xla::InfeedWithToken(*token, *shape);
-
-  return exla::nif::ok(env, exla::nif::make<xla::XlaOp>(env, op));
-}
-
-ERL_NIF_TERM outfeed(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  if (argc != 3) {
-    return exla::nif::error(env, "Bad argument count.");
-  }
-
-  xla::XlaOp* operand;
-  xla::XlaOp* token;
-  xla::Shape* shape_with_layout;
-
-  if (!exla::nif::get<xla::XlaOp>(env, argv[0], operand)) {
-    return exla::nif::error(env, "Unable to get operand.");
-  }
-  if (!exla::nif::get<xla::XlaOp>(env, argv[1], token)) {
-    return exla::nif::error(env, "Unable to get token.");
-  }
-  if (!exla::nif::get<xla::Shape>(env, argv[2], shape_with_layout)) {
-    return exla::nif::error(env, "Unable to get shape.");
-  }
-
-  // TODO(seanmor5): Determine if default config is optimal
-  xla::XlaOp op = xla::OutfeedWithToken(*operand, *token, *shape_with_layout, "");
-
-  return exla::nif::ok(env, exla::nif::make<xla::XlaOp>(env, op));
-}
-
-ERL_NIF_TERM create_token(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  if (argc != 1) {
-    return exla::nif::error(env, "Bad argument count.");
-  }
-
-  xla::XlaBuilder** builder;
-
-  if (!exla::nif::get<xla::XlaBuilder*>(env, argv[0], builder)) {
-    return exla::nif::error(env, "Unable to get builder.");
-  }
-
-  xla::XlaOp token = xla::CreateToken(*builder);
-
-  return exla::nif::ok(env, exla::nif::make<xla::XlaOp>(env, token));
-}
-
-ERL_NIF_TERM optimization_barrier(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  if (argc != 1) {
-    return exla::nif::error(env, "Bad argument count.");
-  }
-
-  xla::XlaOp* operand;
-
-  if (!exla::nif::get<xla::XlaOp>(env, argv[0], operand)) {
-    return exla::nif::error(env, "Unable to get operand.");
-  }
-
-  xla::XlaOp op = xla::OptimizationBarrier(*operand);
-
-  return exla::nif::ok(env, exla::nif::make<xla::XlaOp>(env, op));
 }
 
 ERL_NIF_TERM transfer_to_infeed(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
@@ -2289,7 +717,7 @@ ERL_NIF_TERM compile(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
   }
 
   exla::ExlaClient** client;
-  struct ModuleResource * computation;
+  xla::XlaComputation * computation;
   std::vector<xla::Shape*> argument_layouts;
   xla::ExecutableBuildOptions build_options;
   int num_replicas;
@@ -2300,7 +728,7 @@ ERL_NIF_TERM compile(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
   if (!exla::nif::get<exla::ExlaClient*>(env, argv[0], client)) {
     return exla::nif::error(env, "Unable to get client.");
   }
-  if (!exla::nif::get<struct ModuleResource>(env, argv[1], computation)) {
+  if (!exla::nif::get<xla::XlaComputation>(env, argv[1], computation)) {
     return exla::nif::error(env, "Unable to get computation.");
   }
   if (!exla::nif::get_list<xla::Shape>(env, argv[2], argument_layouts)) {
@@ -2330,7 +758,59 @@ ERL_NIF_TERM compile(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
   }
 
   EXLA_ASSIGN_OR_RETURN_NIF(exla::ExlaExecutable* executable,
-    (*client)->Compile(*(computation->module), argument_layouts, build_options, compile_portable_executable), env);
+    (*client)->Compile(*computation, argument_layouts, build_options, compile_portable_executable), env);
+
+  return exla::nif::ok(env, exla::nif::make<exla::ExlaExecutable*>(env, executable));
+}
+
+ERL_NIF_TERM mlir_compile(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
+  if (argc != 7) {
+    return exla::nif::error(env, "Bad argument count.");
+  }
+
+  exla::ExlaClient** client;
+  exla::MLIRModule** module;
+  std::vector<xla::Shape*> argument_layouts;
+  xla::ExecutableBuildOptions build_options;
+  int num_replicas;
+  int num_partitions;
+  bool use_spmd;
+  int device_id;
+
+  if (!exla::nif::get<exla::ExlaClient*>(env, argv[0], client)) {
+    return exla::nif::error(env, "Unable to get client.");
+  }
+  if (!exla::nif::get<exla::MLIRModule*>(env, argv[1], module)) {
+    return exla::nif::error(env, "Unable to get module.");
+  }
+  if (!exla::nif::get_list<xla::Shape>(env, argv[2], argument_layouts)) {
+    return exla::nif::error(env, "Unable to get argument layouts.");
+  }
+  if (!exla::nif::get(env, argv[3], &num_replicas)) {
+    return exla::nif::error(env, "Unable to get Number of Replicas.");
+  }
+  if (!exla::nif::get(env, argv[4], &num_partitions)) {
+    return exla::nif::error(env, "Unable to get Number of Partitions.");
+  }
+  if (!exla::nif::get(env, argv[5], &use_spmd)) {
+    return exla::nif::error(env, "Unable to get SPMD Partitioning Flag.");
+  }
+  if (!exla::nif::get(env, argv[6], &device_id)) {
+    return exla::nif::error(env, "Unable to get device ID.");
+  }
+
+  build_options.set_num_replicas(num_replicas);
+  build_options.set_num_partitions(num_partitions);
+  build_options.set_use_spmd_partitioning(use_spmd);
+
+  bool compile_portable_executable = false;
+  if (device_id >= 0) {
+    compile_portable_executable = true;
+    build_options.set_device_ordinal(device_id);
+  }
+
+  EXLA_ASSIGN_OR_RETURN_NIF(exla::ExlaExecutable* executable,
+    (*client)->Compile((*module)->module(), argument_layouts, build_options, compile_portable_executable), env);
 
   return exla::nif::ok(env, exla::nif::make<exla::ExlaExecutable*>(env, executable));
 }
@@ -2364,44 +844,11 @@ ERL_NIF_TERM run(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
   return term;
 }
 
-// MLIR Conversion
-
-ERL_NIF_TERM convert_xla_computation_to_mlir_module(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
-  if (argc != 1) {
-    return exla::nif::error(env, "Bad argument count.");
-  }
-
-  xla::XlaComputation* computation;
-
-  if (!exla::nif::get<xla::XlaComputation>(env, argv[0], computation)) {
-    return exla::nif::error(env, "Unable to get computation.");
-  }
-  struct ModuleResource * res =
-    (struct ModuleResource *) enif_alloc_resource(exla::nif::resource_object<struct ModuleResource>::type, sizeof(ModuleResource));
-
-  new (&res->context) mlir::MLIRContext();
-  new (&res->module) mlir::OwningOpRef<mlir::ModuleOp>(mlir::ModuleOp::create(mlir::UnknownLoc::get(&res->context)));
-
-  (res->context).loadDialect<mlir::func::FuncDialect>();
-  (res->context).loadDialect<mlir::mhlo::MhloDialect>();
-
-  xla::Status status = ConvertHloToMlirHlo(*(res->module), &(computation->proto()), true);
-
-  ERL_NIF_TERM term = enif_make_resource(env, res);
-  enif_release_resource(res);
-
-  if (!status.ok()) {
-    return exla::nif::error(env, status.message().data());
-  } else {
-    return exla::nif::ok(env, term);
-  }
-}
-
 // Logging Functions
 
 ERL_NIF_TERM start_log_sink(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[]) {
   if (argc != 1) {
-    return exla::nif::make(env, "Bad argument count.");
+    return exla::nif::error(env, "Bad argument count.");
   }
 
   ErlNifPid logger_pid;
@@ -2423,6 +870,14 @@ ERL_NIF_TERM start_log_sink(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[])
 }
 
 static ErlNifFunc exla_funcs[] = {
+  // MLIR Builder
+  {"new_mlir_module", 0, new_mlir_module},
+  {"create_mlir_function", 4, create_mlir_function},
+  {"get_mlir_function_arguments", 1, get_mlir_function_arguments},
+  {"mlir_add", 3, mlir_add},
+  {"mlir_subtract", 3, mlir_subtract},
+  {"mlir_build", 2, mlir_build},
+  {"dump_mlir_module", 1, dump_mlir_module},
   // XlaBuilder
   {"new_builder", 1, new_builder},
   {"create_sub_builder", 2, create_sub_builder},
@@ -2437,6 +892,7 @@ static ErlNifFunc exla_funcs[] = {
   {"get_device_count", 1, get_device_count},
   {"get_supported_platforms", 0, get_supported_platforms},
   {"compile", 7, compile, ERL_NIF_DIRTY_JOB_CPU_BOUND},
+  {"mlir_compile", 7, mlir_compile, ERL_NIF_DIRTY_JOB_CPU_BOUND},
   // ExlaBuffer
   {"binary_to_device_mem", 4, binary_to_device_mem, ERL_NIF_DIRTY_JOB_IO_BOUND},
   {"read_device_mem", 2, read_device_mem, ERL_NIF_DIRTY_JOB_IO_BOUND},
@@ -2522,10 +978,12 @@ static ErlNifFunc exla_funcs[] = {
   // Tuples
   {"tuple", 2, tuple},
   {"get_tuple_element", 2, get_tuple_element},
-  // Conditionals
+  // Control Flow
   {"conditional", 5, conditional_if},
   {"conditional", 3, conditional_multi},
   {"select", 3, select},
+  {"while", 3, while_loop},
+  {"call", 3, call},
   // Slicing
   {"slice", 4, slice},
   {"dynamic_slice", 3, dynamic_slice},
@@ -2542,8 +1000,6 @@ static ErlNifFunc exla_funcs[] = {
   {"select_and_scatter", 8, select_and_scatter},
   {"scatter", 8, scatter},
   {"map", 4, map},
-  {"while", 3, while_loop},
-  {"call", 3, call},
   // Shape/Type Manipulation
   {"broadcast_in_dim", 3, broadcast_in_dim},
   {"reshape", 2, reshape},
@@ -2575,9 +1031,7 @@ static ErlNifFunc exla_funcs[] = {
   // Special
   {"optimization_barrier", 1, optimization_barrier},
   // Log Sink
-  {"start_log_sink", 1, start_log_sink},
-  // MLIR
-  {"convert_xla_computation_to_mlir_module", 1, convert_xla_computation_to_mlir_module}
+  {"start_log_sink", 1, start_log_sink}
 };
 
 ERL_NIF_INIT(Elixir.EXLA.NIF, exla_funcs, &load, NULL, NULL, NULL);
