+++ exla/lib/exla/client.ex
@@ -172,7 +172,12 @@ defmodule EXLA.Client do
 
         :metal ->
           # TODO: Is this really where/how we want to do this?
-          :ok = EXLA.NIF.load_pjrt_plugin("METAL", "/opt/homebrew/lib/python3.10/site-packages/jax_plugins/metal_plugin/pjrt_plugin_metal_14.dylib")
+          :ok =
+            EXLA.NIF.load_pjrt_plugin(
+              "METAL",
+              "/opt/homebrew/lib/python3.10/site-packages/jax_plugins/metal_plugin/pjrt_plugin_metal_14.dylib"
+            )
+
           EXLA.NIF.get_c_api_client("METAL")
 
         _ ->
