+++ exla/test/exla/mlir/executable_test.exs
@@ -1,0 +1,6 @@
+defmodule EXLA.MLIR.ExecutableTest do
+  use ExUnit.Case, async: true
+
+  alias EXLA.{BinaryBuffer, DeviceBuffer, Executable}
+  import EXLAHelpers
+end
